workflow/requirements.txt
python-socketio
opencv-python
python-dotenv
uvicorn
inference